REF: move Poisson dispersion tests to discrete
