+++ statsmodels/discrete/diagnostic.py
@@ -13,15 +13,12 @@ import numpy as np
 
 from statsmodels.tools.decorators import cache_readonly
 
-from statsmodels.stats._diagnostic_other import (
-    dispersion_poisson,
-    # dispersion_poisson_generic,
-    )
-
 from statsmodels.stats.diagnostic_gen import (
     test_chisquare_binning
     )
 from statsmodels.discrete._diagnostics_count import (
+    dispersion_poisson,
+    # dispersion_poisson_generic,
     test_poisson_zeroinflation_jh,
     test_poisson_zeroinflation_broek,
     test_poisson_zeros,
@@ -37,7 +34,10 @@ class CountDiagnostic:
 
     Parameters
     ----------
-    results : PoissonResults instance
+    results : Results instance of a count model.
+    y_max : int
+        Largest count to include when computing predicted probabilities for
+        counts. Default is the largest observed count.
 
     """
 
@@ -48,7 +48,7 @@ class CountDiagnostic:
     @cache_readonly
     def probs_predicted(self):
         if self.y_max is not None:
-            kwds = {"y_values": np.arange(self.y_max)}
+            kwds = {"y_values": np.arange(self.y_max + 1)}
         else:
             kwds = {}
         return self.results.predict(which="prob", **kwds)
