+++ statsmodels/stats/_diagnostic_other.py
@@ -659,159 +659,6 @@ def lm_robust_reparameterized(score, params_deriv, score_deriv, cov_score):
     return lm_stat, pval
 
 
-def dispersion_poisson(results):
-    """Score/LM type tests for Poisson variance assumptions
-
-    Null Hypothesis is
-
-    H0: var(y) = E(y) and assuming E(y) is correctly specified
-    H1: var(y) ~= E(y)
-
-    The tests are based on the constrained model, i.e. the Poisson model.
-    The tests differ in their assumed alternatives, and in their maintained
-    assumptions.
-
-    Parameters
-    ----------
-    results : Poisson results instance
-        This can be a results instance for either a discrete Poisson or a GLM
-        with family Poisson.
-
-    Returns
-    -------
-    res : ndarray, shape (7, 2)
-       each row contains the test statistic and p-value for one of the 7 tests
-       computed here.
-    description : 2-D list of strings
-       Each test has two strings a descriptive name and a string for the
-       alternative hypothesis.
-    """
-
-    if hasattr(results, '_results'):
-        results = results._results
-
-    endog = results.model.endog
-    nobs = endog.shape[0]   #TODO: use attribute, may need to be added
-    fitted = results.predict()
-    #fitted = results.fittedvalues  # discrete has linear prediction
-    #this assumes Poisson
-    resid2 = results.resid_response**2
-    var_resid_endog = (resid2 - endog)
-    var_resid_fitted = (resid2 - fitted)
-    std1 = np.sqrt(2 * (fitted**2).sum())
-
-    var_resid_endog_sum = var_resid_endog.sum()
-    dean_a = var_resid_fitted.sum() / std1
-    dean_b = var_resid_endog_sum / std1
-    dean_c = (var_resid_endog / fitted).sum() / np.sqrt(2 * nobs)
-
-    pval_dean_a = 2 * stats.norm.sf(np.abs(dean_a))
-    pval_dean_b = 2 * stats.norm.sf(np.abs(dean_b))
-    pval_dean_c = 2 * stats.norm.sf(np.abs(dean_c))
-
-    results_all = [[dean_a, pval_dean_a],
-                   [dean_b, pval_dean_b],
-                   [dean_c, pval_dean_c]]
-    description = [['Dean A', 'mu (1 + a mu)'],
-                   ['Dean B', 'mu (1 + a mu)'],
-                   ['Dean C', 'mu (1 + a)']]
-
-    # Cameron Trived auxiliary regression page 78 count book 1989
-    endog_v = var_resid_endog / fitted
-    res_ols_nb2 = OLS(endog_v, fitted).fit(use_t=False)
-    stat_ols_nb2 = res_ols_nb2.tvalues[0]
-    pval_ols_nb2 = res_ols_nb2.pvalues[0]
-    results_all.append([stat_ols_nb2, pval_ols_nb2])
-    description.append(['CT nb2', 'mu (1 + a mu)'])
-
-    res_ols_nb1 = OLS(endog_v, fitted).fit(use_t=False)
-    stat_ols_nb1 = res_ols_nb1.tvalues[0]
-    pval_ols_nb1 = res_ols_nb1.pvalues[0]
-    results_all.append([stat_ols_nb1, pval_ols_nb1])
-    description.append(['CT nb1', 'mu (1 + a)'])
-
-    endog_v = var_resid_endog / fitted
-    res_ols_nb2 = OLS(endog_v, fitted).fit(cov_type='HC3', use_t=False)
-    stat_ols_hc1_nb2 = res_ols_nb2.tvalues[0]
-    pval_ols_hc1_nb2 = res_ols_nb2.pvalues[0]
-    results_all.append([stat_ols_hc1_nb2, pval_ols_hc1_nb2])
-    description.append(['CT nb2 HC3', 'mu (1 + a mu)'])
-
-    res_ols_nb1 = OLS(endog_v, np.ones(len(endog_v))).fit(cov_type='HC3',
-                                                          use_t=False)
-    stat_ols_hc1_nb1 = res_ols_nb1.tvalues[0]
-    pval_ols_hc1_nb1 = res_ols_nb1.pvalues[0]
-    results_all.append([stat_ols_hc1_nb1, pval_ols_hc1_nb1])
-    description.append(['CT nb1 HC3', 'mu (1 + a)'])
-
-    return np.array(results_all), description
-
-
-def dispersion_poisson_generic(results, exog_new_test, exog_new_control=None,
-                               include_score=False, use_endog=True,
-                               cov_type='HC3', cov_kwds=None, use_t=False):
-    """A variable addition test for the variance function
-
-    This uses an artificial regression to calculate a variant of an LM or
-    generalized score test for the specification of the variance assumption
-    in a Poisson model. The performed test is a Wald test on the coefficients
-    of the `exog_new_test`.
-
-    Warning: insufficiently tested, especially for options
-    """
-
-    if hasattr(results, '_results'):
-        results = results._results
-
-    endog = results.model.endog
-    nobs = endog.shape[0]   #TODO: use attribute, may need to be added
-    # fitted = results.fittedvalues  # generic has linpred as fittedvalues
-    fitted = results.predict()
-    resid2 = results.resid_response**2
-    #the following assumes Poisson
-    if use_endog:
-        var_resid = (resid2 - endog)
-    else:
-        var_resid = (resid2 - fitted)
-
-    endog_v = var_resid / fitted
-
-    k_constraints = exog_new_test.shape[1]
-    ex_list = [exog_new_test]
-    if include_score:
-        score_obs = results.model.score_obs(results.params)
-        ex_list.append(score_obs)
-
-    if exog_new_control is not None:
-        ex_list.append(score_obs)
-
-    if len(ex_list) > 1:
-        ex = np.column_stack(ex_list)
-        use_wald = True
-    else:
-        ex = ex_list[0]  # no control variables in exog
-        use_wald = False
-
-    res_ols = OLS(endog_v, ex).fit(cov_type=cov_type, cov_kwds=cov_kwds,
-                                   use_t=use_t)
-
-    if use_wald:
-        # we have controls and need to test coefficients
-        k_vars = ex.shape[1]
-        constraints = np.eye(k_constraints, k_vars)
-        ht = res_ols.wald_test(constraints)
-        stat_ols = ht.statistic
-        pval_ols = ht.pvalue
-    else:
-        # we do not have controls and can use overall fit
-        nobs = endog_v.shape[0]
-        rsquared_noncentered = 1 - res_ols.ssr/res_ols.uncentered_tss
-        stat_ols = nobs * rsquared_noncentered
-        pval_ols = stats.chi2.sf(stat_ols, k_constraints)
-
-    return stat_ols, pval_ols
-
-
 def conditional_moment_test_generic(mom_test, mom_test_deriv,
                                     mom_incl, mom_incl_deriv,
                                     var_mom_all=None,
